add: split check_and_inference expr_stmt function
