+++ src/type_parser/check_and_inference/mod.rs
@@ -1,3 +1,4 @@
 pub mod check_and_inference;
-mod variable_declaration;
 mod _struct;
+mod variable_declaration;
+mod expr_stmt;
